Fix a compile warning (oops).
It is uncouth to use an int to iterate up to a size_t.

+++ crawl-ref/source/rng.cc
@@ -30,7 +30,7 @@ void seed_rng(uint32_t* seed_key, size_t num_keys)
 
     // for std::random_shuffle()
     uint32_t oneseed = 0;
-    for (int i = 0; i < num_keys; ++i)
+    for (size_t i = 0; i < num_keys; ++i)
         oneseed += seed_key[i];
 
     srand(oneseed);
